拆分多个类型实现不同的功能
减少单个类包含太多逻辑
